requirements.txt
tensorflow>=2.8.0
numpy>=1.19.5
opencv-python>=4.5.5
scikit-image>=0.19.0
scikit-learn>=1.0.2
matplotlib>=3.5.1
seaborn>=0.11.2
pandas>=1.3.5
scipy>=1.7.3
tqdm>=4.62.3
pillow>=9.0.0
pyyaml>=6.0
h5py>=3.6.0
